Changed ints to longs (pertaining to totalLineCount)

+++ FileSplitter/FIlenameGenerator.cs
@@ -10,13 +10,13 @@ namespace FileSplitter {
 
         public static string FilenameRoot { get; set; }
         public static string Extension { get; set; }
-        public static int TotalLines { get; set; }
+        public static long TotalLines { get; set; }
         public static int MaxLinesPerSplitFile { get; set; }
         public static int TotalSplitFiles { get; set; }
         public static int DigitsInFileCount { get; set; }
         public static int DigitsInLineCount { get; set; }
 
-        public FilenameGenerator(string path, int totalLines, int linesToSplit) {
+        public FilenameGenerator(string path, long totalLines, int linesToSplit) {
             FilenameRoot = Path.GetFileNameWithoutExtension(path);
             Extension = Path.GetExtension(path);
             TotalLines = totalLines;
@@ -29,12 +29,12 @@ namespace FileSplitter {
         public string GenerateFilename(int currentFileCount = 1, int startingLineNumber = 1) {
             string fileCountDecimalFormat = MathHelpers.GetNumericDecimalFormatString(TotalSplitFiles);
             string lineCountDecimalFormat = MathHelpers.GetNumericDecimalFormatString(TotalLines);
-            int endingLineNumber = CalculateEndingLineNumber(startingLineNumber);
+            long endingLineNumber = CalculateEndingLineNumber(startingLineNumber);
             return $"{FilenameRoot}_{currentFileCount.ToString(fileCountDecimalFormat)}_{startingLineNumber.ToString(lineCountDecimalFormat)}-{endingLineNumber.ToString(lineCountDecimalFormat)}{Extension}";
         }
 
-        private static int CalculateEndingLineNumber(int startingLineNumber) {
-            int endingLineNumber = startingLineNumber + MaxLinesPerSplitFile - 1;
+        private static long CalculateEndingLineNumber(int startingLineNumber) {
+            long endingLineNumber = startingLineNumber + MaxLinesPerSplitFile - 1;
             if (endingLineNumber > TotalLines) {
                 endingLineNumber = TotalLines;
             }
